greater or lesser fix

+++ returnFloatingPoint/main.cpp
@@ -10,7 +10,7 @@
 using namespace std;
 
 // extern keyword is used to declare a variable that is defined in another file.
-extern "C" int floating_point_processor();
+extern "C" double floating_point_processor();
 
 // this is the driver function for the program
 int main()
